pyproject.toml
[project]
name = "medical-exam-prep-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "audioop-lts==0.2.1",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.175.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.24.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.1",
    "grpcio-status==1.71.2",
    "gtts==2.5.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "markupsafe==3.0.2",
    "narwhals==1.45.0",
    "numpy==2.3.1",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.3.0",
    "plotly>=6.2.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyaudio>=0.2.14",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "speechrecognition==3.14.3",
    "standard-aifc==3.13.0",
    "standard-chunk==3.13.0",
    "streamlit==1.46.1",
    "streamlit-player==0.1.5",
    "tenacity==8.5.0",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "websockets==15.0.1",
]
